requirements.txt
ffmpeg-python==0.2.0
fortune-python==1.1.1
future==1.0.0
